Pulled out get_grib_input to live under tools.
